Move analysis entry points to analysis module
