+++ external-crates/move/crates/move-analyzer/src/analysis/mod.rs
@@ -1,5 +1,98 @@
 // Copyright (c) The Move Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::{
+    compiler_info::CompilerInfo,
+    symbols::{
+        cursor::CursorContext,
+        types::{CompiledPkgInfo, SymbolsComputationData, UseDefMap},
+    },
+    utils::expansion_mod_ident_to_map_key,
+};
+
+use im::ordmap::OrdMap;
+use lsp_types::Position;
+use std::collections::BTreeMap;
+
+use move_compiler::{
+    expansion::ast::ModuleIdent,
+    parser::ast as P,
+    shared::{NamedAddressMap, files::MappedFiles, unique_map::UniqueMap},
+    typing::{ast::ModuleDefinition, visitor::TypingVisitorContext},
+};
+
 pub mod parsing_analysis;
 pub mod typing_analysis;
+
+/// Run parsing analysis for either main program or dependencies
+pub fn run_parsing_analysis(
+    computation_data: &mut SymbolsComputationData,
+    compiled_pkg_info: &CompiledPkgInfo,
+    cursor_context: Option<&mut CursorContext>,
+    parsed_program: &P::Program,
+) {
+    let mut parsing_symbolicator = parsing_analysis::ParsingAnalysisContext {
+        mod_outer_defs: &mut computation_data.mod_outer_defs,
+        files: &compiled_pkg_info.mapped_files,
+        references: &mut computation_data.references,
+        def_info: &mut computation_data.def_info,
+        use_defs: UseDefMap::new(),
+        current_mod_ident_str: None,
+        alias_lengths: BTreeMap::new(),
+        pkg_addresses: &NamedAddressMap::new(),
+        cursor: cursor_context,
+    };
+
+    parsing_symbolicator.prog_symbols(
+        parsed_program,
+        &mut computation_data.mod_use_defs,
+        &mut computation_data.mod_to_alias_lengths,
+    );
+}
+
+/// Run typing analysis for either main program or dependencies
+pub fn run_typing_analysis(
+    mut computation_data: SymbolsComputationData,
+    mapped_files: &MappedFiles,
+    compiler_info: &mut CompilerInfo,
+    typed_program_modules: &UniqueMap<ModuleIdent, ModuleDefinition>,
+) -> SymbolsComputationData {
+    let mut typing_symbolicator = typing_analysis::TypingAnalysisContext {
+        mod_outer_defs: &mut computation_data.mod_outer_defs,
+        files: mapped_files,
+        references: &mut computation_data.references,
+        def_info: &mut computation_data.def_info,
+        use_defs: UseDefMap::new(),
+        current_mod_ident_str: None,
+        alias_lengths: &BTreeMap::new(),
+        traverse_only: false,
+        compiler_info,
+        type_params: BTreeMap::new(),
+        expression_scope: OrdMap::new(),
+    };
+
+    process_typed_modules(
+        typed_program_modules,
+        &computation_data.mod_to_alias_lengths,
+        &mut typing_symbolicator,
+        &mut computation_data.mod_use_defs,
+    );
+    computation_data
+}
+
+fn process_typed_modules<'a>(
+    typed_modules: &UniqueMap<ModuleIdent, ModuleDefinition>,
+    mod_to_alias_lengths: &'a BTreeMap<String, BTreeMap<Position, usize>>,
+    typing_symbolicator: &mut typing_analysis::TypingAnalysisContext<'a>,
+    mod_use_defs: &mut BTreeMap<String, UseDefMap>,
+) {
+    for (module_ident, module_def) in typed_modules.key_cloned_iter() {
+        let mod_ident_str = expansion_mod_ident_to_map_key(&module_ident.value);
+        typing_symbolicator.use_defs = mod_use_defs.remove(&mod_ident_str).unwrap();
+        typing_symbolicator.alias_lengths = mod_to_alias_lengths.get(&mod_ident_str).unwrap();
+        typing_symbolicator.visit_module(module_ident, module_def);
+
+        let use_defs = std::mem::replace(&mut typing_symbolicator.use_defs, UseDefMap::new());
+        mod_use_defs.insert(mod_ident_str, use_defs);
+    }
+}
